googlebigquery/requirements.txt
google-cloud==0.27.0
gcs-oauth2-boto-plugin==1.14